Improve linting setup, stricter TS and GraphQL rules
- Typescript-eslint config now includes strict rules. Minor changes throughout ui package
- Added `@graphql-eslint/eslint-plugin` to lint GraphQL files
   - Added in schema package to lint schema files, minor changes for comments which now appear in the types
   - Added in ui package to lint operations. Also warns for deprecated fields and types (main purpose I started on this). Also caught a mistake I made from unused variables
- Replace another deprecated GraphQL query we were using. Glad to see the rule works 🎉

Unfortunately we can't lint the operations for navigate-server and navigate-configs at the same time. This is because it tries to merge the schemas together, which brings some incompatible names and type conflicts. Linting for the `navigate-configs` operations is commented out, so it's easy to switch if needed. The operations are also much more straightforward.